Update patch set 1
Patch Set 1:

This change isn't really needed, if you are using the docker-compose.yaml in sdnc/oam/installation/src/main/yaml to start containers.  The entry point defined in that docker-compose is a shell script, which performs the chmod.  However there is no harm doing it in Dockerfile as well.

Patch-set: 1

Former-commit-id: 03400ce85badbbb02d6b96b67508b41ac2daecee