Update patch set 1
Patch Set 1:

It's worth noting that some of the DGs, for both the Generic API and the VNF API (xml and json versions) use old "com.att" resource locations.  IpAddressTool is one example where I'm sure it should be updated.  

The EIPAMPlugin is another example, but I'm unclear whether this is intentional or not.

Patch-set: 1

Former-commit-id: bf384d72c7db87fb1e67a8133e7748029ec6fdc7