Update patch set 1
Patch Set 1:

Note: This patch assumes that for deactivate SO will provide the siteId, it can happen that SO will rather provide some recordId; this recordId would have been provided by SDNC to SO after storing data to A&AI during activate.

Later will followup with SO team to get this information and will add a new commit if required.

Patch-set: 1

Former-commit-id: 124abbe5709de207431148eebe7edef4b17814fe