Update patch set 1
Patch Set 1:

Sorry Brian, I didn't see the email until after I merged that mentioned to wait on this until testing. Good practice in these cases is to put "Do Not Merge" in summary and -1 if something is a test change. Then remove them when patch is considered ready. I'll just submit some updates post testing if we find issues with this, but first reading seemed ok to me.

Patch-set: 1

Former-commit-id: c6bab0babe3f55e67ecdc17d1daf6e57bb14ada3