Added external_key column to IPAM_IP_ASSIGNMENT
external_key column is to store a keu uniquely identifying
the resource to a service, both within the ONAP and the
external system. That key will be set prior assignment,
and we be used during unassignment, to avoid having to
correlate to many values. The external_key *must* be unique,
at least per service. Hence lookup using both service-instance-id
and external_key will be used during unassign process.

Change-Id: I67691fc32b46bb6cbacd66402f5fda6456cd7e06
Issue-ID: CCSDK-463
Signed-off-by: Alexis de Talhouët <adetalhouet89@gmail.com>

Former-commit-id: b3cc6305655fee9a85d8ac42488e371b741a2f61