Update patch set 1
Patch Set 1:

The easiest thing to do is just to use dgbuilder to load the JSON version, and then export the XML version.  One enhancement I'd like to make in the future is to make that conversion automatic (i.e. to enhance the compiler so that it can convert JSON to XML, and then compile XML) but haven't had chance to work on that yet.

Patch-set: 1
Label: Code-Review=0

Former-commit-id: 1f0275c93259d68b64b35949dda8f6458ff3b016