Update patch set 2
Patch Set 2: Code-Review+1

We might want to create 2 dockers - one integrated with AAF, and one not.  That way, we can use the standalone version for development testing (where we might not have access to AAF). The existing one (sdnc) should be integrated with AAF as you.ve done, but we might want to add a new "sdncdev" directory under installation that creates a docker (sdncdev) not integrated with AAF that we can use for local testing

Patch-set: 2
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1

Former-commit-id: a9fc180bdef865e7c05924cc51fbfa22d46494c3