Update patch set 1
Patch Set 1: Code-Review+1

Looks good.  I didn't recall this the last time we talked, but we'd already reserved a nodeport in OOM ( 30267 ) for the SDNC pod to expose 8443 back in Beijing.  So once this is in, new images should be testable in an OOM deployment.

Patch-set: 1
Reviewer: Gerrit User 659 <659@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1

Former-commit-id: 732928a2f117d70bb9df4b54439aa31fde2d90cd