Drop CDS tables only if they exist
During the initialization of SDNC, SDNC tries to install all
the SQL dump provided. While processing the one for CDS, it
was failing to execute the first statement as the table is not
existent; hence was exiting that dump and processing with the
next one. In order to make sure the process creates CDS tables,
make the DROP statement work even if table doesn't exist.

Change-Id: Ifa99b8e1a13ac3828828e8b2a9f9674556f84ced
Issue-ID: CCSDK-682
Signed-off-by: Alexis de Talhouët <adetalhouet89@gmail.com>

Former-commit-id: d9dbcdbe0cb870b558ef0415e5c33fef01413128