Update patch set 1
Patch Set 1: -Code-Review

Hi Dan, this is required for CCVPN use case, which is based on SO E2E service instance flow wherein SO is not sending the vnf-id, rather it's expected from SDNC to generate vnf-id as part of resource creation followed by creation of respective record in AAI. SO obtains vnf-id from SDNC response message.

I think this change was merged earlier, but later was overwritten by some other patch. This patch aims to make the SDNC implementation flexible to allow SDNC generate vnf-id for assign/create operation rather then mandating SO to provide it.

This problem is identified during the on-going CCVPN integration test.

Patch-set: 1

Former-commit-id: 8ee913e92bfd58d5bf10cf5b43c0f900e0462972