Update patch set 5
Patch Set 5: Code-Review-2

This change is not correct.

The SDNC ansible container is intended to be the CCSDK ansible container base, with SDNC specific playbooks added on top.  This change would replace that with no reuse across SDNC and APPC, and with the SDNC playbooks removed.

These do look like good changes that should be made to the ccsdk ansible container (in ccsdk/distribution), and I'd be happy to merge those changes there.  That way both SDNC and APPC could benefit.

Patch-set: 5
Label: Code-Review=-2

Former-commit-id: 14377d257d8593a91bacd5f8e49c362d21b06b59