Update patch set 1
Patch Set 1:

If I add it in the startODL.sh script it will be added to the end of the feature line and not beside the mdsal feature setup. This was the idea behind. Please give indication if I should add it to startODL.sh or let it in the Docerfile.

Patch-set: 1

Former-commit-id: 43b653dfab516a07b5d54ea8886ef9ebbbcf723e