Update patch set 1
Patch Set 1:

The intention is only to install odl-mdsal-clustering if clustering is enabled (i.e. if the number of replicas is > 1).  That's what we were doing before.  So we can't bake it into the Dockerfile - we have to add it to the featureBoot in the startODL.sh, just like you're doing for odl-jolokia

Patch-set: 1
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0

Former-commit-id: a9e154679b515d5b6501b6eb73b4eb30c078bd1b