Update patch set 1
Patch Set 1:

The Yang models of upgrade-pre-check and upgrade-post-check have a mandatory payload field in their output, and the 'outputPayload' deleted in this patch is used to set the payload field of the output. This payload is used for PNF software upgrade use case.

The 'outputPayload' is got from the host[output] field of results of Ansible server, and the host[output] is got from Ansible playbook command as: 
local_action: copy content="{{precheck_result.stdout}}" dest="{{inventory_dir}}/{{inventory_hostname}}_results.txt"

If the Ansible playbook doesn't have a command that write contents to {{inventory_dir}}/{{inventory_hostname}}_results.txt, the 'outputPayload' is null and the outputs of upgrade-pre-check and upgrade-post-check don't have the payload filed. So, in my option, the 'outputPayload' in the DGs doesn't have a negative impact.

Therefore, could you reserve the 'outputPayload' in the DGs for the PNF software upgrade use case?

Patch-set: 1
Reviewer: Gerrit User 5302 <5302@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0

Former-commit-id: 6e1e6c986081cc5a52fbdf8264551f5df8138975