Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi, I don't know how your build process works so if something is missing here then pls just let me know or take over this commit and integrate it in a better way.
> 
> The goal is to remove any runtime dependencies for docker container which the current version violates. Downloading anything after container startup is truly anti-pattern and it kills the purpose to use containers in the first place.
> 
> Thanks

Thanks - this should be fine.

FYI - the reason we ran npm in the startup script is that originally it had been failing when we ran in Jenkins as part of the build.  We later discovered the root cause of that error and corrected it, but didn't remember to move npm back (note: admin portal is disabled in Dublin anyway due to other issues).  Thanks for moving it back where it belongs!

Patch-set: 1

Former-commit-id: a2075b99ba6ae5d08b51ce09af72ae5117b56698