Proposal to remove OSGi dependencies from the SDNC project
Dependencies on the OSGi frameworks and libraries are removed
by integrating the CCSDK and SDNC project with the lighty.io.
It's a toolkit that allows to use ODL services (in this case
core services and the Restconf) without the dependency
on the Karaf framework and the Blueprint DI.

In this change are created the lighty.io modules which
initialize and expose same services as the Blueprint DI in
the blueprint.xml files.

More info about the lighty.io - https://lighty.io

Signed-off-by: Samuel Kontris <samuel.kontris@pantheon.tech>
Issue-ID: SDNC-803
Change-Id: Id0582cd2a264e2538d5024b32d45e72058079e31
Signed-off-by: Samuel Kontris <samuel.kontris@pantheon.tech>

Former-commit-id: 7d4faf5d2d0f07fe2be09aff18818652f5661c12