Update patch set 2
Patch Set 2: Code-Review+1

Note: in kubernetes env, there is a separate copy of startODLsh (oom/kubernetes/sdnc/resources/config/bin/startODL.sh) ... so please also create a change in the oom project to apply these changes to the startODL.sh there as well.

Also, note that we're required to disable https in the case that the helm variable global.aafEnabled is set to false.  That doesn't apply to the version of startODL.sh here, since this version is only ever used in a non-kubernetes env (e.g. docker-compose).

Patch-set: 2
Label: Code-Review=+1

Former-commit-id: 1ddb68fb5dd3d2ee16271160f5ecdc7f1aaefd35