Update patch set 1
Patch Set 1:

Thanks Herbert for your comment.

We have another related commit in ccsdk/distribution that adds the a1-adapter-northbound-installer to odlsli-alpine image:

https://gerrit.onap.org/r/c/ccsdk/distribution/+/111205

I think this commit should be merged before we can see the a1-adapter working in sdnc-image (since the feature will not be a part of odlsli-alpine image otherwise). We got a comment from Dan regarding the versioning that we are trying to fix. So, we will get back to this commit later.

Regarding your question, we are trying to converge SDNR's a1-adapter & OSC's a1-adapter into ccsdk/oran repo (that's where our new bundle is defined). Once that's done, we plan to deprecate the SDNR's a1-adapter.

Hope that makes it clear.

> Patch Set 1: Code-Review-1
> 
> Because of my comment before, I give a -1. This needs to be solved and answered first. 
> Chat contact is #sdnr Rocketchat.

Patch-set: 1