Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Has this been tested?

That is to say, have you tested this change both in standalone and in OOM environment?

In OOM, we're still mounting a separate version of both scripts.  I'd like to change that and use this combined script instead, but first I'd like to confirm it's been tested in the OOM environment.

Patch-set: 3
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>

Former-commit-id: 0b736cbdc3f48812f5a4ada83b64c67404934c82