Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Is this still needed?  Looks like only a spacing change, so maybe fixed in another review?

This is no longer required as the sdnc builds are successful now after yesterday's fix in data-provider. I think Herbert created it to trigger the oam build on demand to verify if the builds are successful instead of waiting for the Jenkins job to do it once a day.

Patch-set: 1

Former-commit-id: 49a06987cab872886de82f2c917d1e7e58704748