Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> SDNC/CCSDK Kohn are in preparation for M3 = final code submission. So mainly bug fixes are expected. Initial submission sounds to me that it is a bigger task that should go better to ONAP London. 
> What is the plan here?

Thanks very much for your comment. We promissed and planed to support this use-case in Kohn release. We will do anything needed to finalize the code on time. This submission although may look large, but in-fact is mostly related to some DG changes which are mostly use-case dependent and have no impact on other use-cases.
Also although my submission says "initial submission ...", but pretty much all the files I plan to submit for this use-case is included in this submission and from now mostly bug fixes will be submitted for these files.

Patch-set: 1

Former-commit-id: 42bfade439c8c74cd7c260945b485adfcdae491c